Added a check to force the secret_key configuration.
This is to catch typos with potential security implications. All the
other missing variables have safe default, but adding a test of the
availiblity of the secret_key for each of its usage would cost too much
CPU. This is why we do it once at the start of the process.